Unify AddToListCommand variants
