--- src/main/java/com/elsria/commands/ToDoCommand.java
@@ -1,24 +1,0 @@
-package com.elsria.commands;
-
-import com.elsria.core.ApplicationContext;
-import com.elsria.task.Task;
-import com.elsria.task.ToDoTask;
-
-/**
- * TODO:Scrap this class
- */
-public class ToDoCommand extends AddToListCommand {
-
-    public ToDoCommand(ApplicationContext context, CommandRequest request) {
-        super(context, request);
-    }
-
-    @Override
-    public Task createTask(String args) {
-        if (args.isEmpty()) {
-            super.errorMessage = "Erm... so what task?";
-            return null;
-        }
-        return new ToDoTask(args);
-    }
-}
